pyproject.toml
[tool.poetry]
name = "emo-classifier"
version = "0.1.0"
description = "Classify whether a song is emo or not emo using Spotify features or raw audio"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = ">=1.0.0"
spotipy = ">=2.23.0"
requests = ">=2.31.0"
librosa = ">=0.10.0"
soundfile = ">=0.12.0"
numpy = ">=1.24.0"
scipy = ">=1.10.0"
scikit-learn = ">=1.3.0"
xgboost = ">=2.0.0"
pandas = ">=2.0.0"
fastapi = ">=0.104.0"
uvicorn = {extras = ["standard"], version = ">=0.24.0"}
joblib = "^1.5.3"
python-multipart = "^0.0.22"

[tool.poetry.group.dev.dependencies]
# PyInstaller supports Python 3.8-3.14; use build_portable.py if on 3.14+
pyinstaller = {version = ">=6.0", python = ">=3.10,<3.14"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
